Return empty BIL image on non intersecting box
Return an empty image when the requested box does not intersect
with the sources coverage. It used to throw an error and thus return an
xml file with the error in spite of an image.

Make it consistent with the behaviour of other image formats:
Other formats like tiff or png are returning blank images.

Make it consistent with partial intersection behaviour:
When the requested box is not completly covered by the sources, the
image is completed with blank pixels.

Note that 'blank' is hexa '0080'.