[GEOS-5993] Missing ElectronicMailAddress in WFS capabilities
In the WFS capabilities there were missing two important fields that are
present in the XSD. Those fields are ElectronicMailAddress and
DeliveryPoint.

This commit adds those two fields, but disregards the fact that there
can be more than one of both fields. However this patch could be adapted
to create a List or Array of both, so they are more faithful to the XSD.