[GEOS-6655] Add text/xml MIME-types to WMS
This commit adds "text/xml" and "text/xml; subtype=gml/3.1.1" MIME-types
to WMS 1.1.1 and 1.3.0. These MIME-types are similar to the already
existing GML2 and GML3.1.1 MIME-types that are already present in
GeoServer.

The main reasons for adding those two MIME-types is 1) compatibility
with the OGC WMS standard, which defines that text/xml should be
expected on a textual output, and 2) supporting the Dutch national
profile for maps. Both documents are linked on JIRA.