GEOS-6688 - WCS 1.1.1 GetCoverage request fails
Edited the WCS 1.1.1 class to let cubic (as bicubic) and linear (as
bilinear) be used as valid interpolation methods. Those were disabled in
code and GeoServer would serve an Exception when trying to access those
interpolations.

I've also added prelimenary support for bicubic and bilinear as keywords
and made a very small performance improvement on GetCoverage requests.