[GEOS-6688-3] Fix for nearest neighbor interpolation (with tests).
Apparently my earlier fix wasn't sufficient and didn't work all that well. That's why I changed the order of the check in the String which makes the check more stable. Also I've added an extra test for RangeSubset with interpolation in GetCoverage (where it belongs).

Finally I've renamed Nearest Neighbour in the web interface to Nearest Neighbor which seems to fix the web interface from not displaying this interpolation method when loading a new layer. Now it works properly. I've also addressed my potential concerns about this on the developers mailing list.

(Resubmitted from #847)