[GEOS-6841] Fix for invalid email address on the front page
As reported in https://jira.codehaus.org/browse/GEOS-6841, this delivers a quick fix for the issue by checking if the mail address is equals to null. If it is this will be replaced by geoserver@example.org on the homepage.

I specifically choose example org because it is an IANA reserved domain for test purposes and can be used without prior consent.