Fix a NPE when changing a SQL view for a layer.
ResourceConfigurationPage#updateResource can be called with target==null.
So we need to check for it.

A NPE was always raised when the user edits an existing SQL view and was hitting
the save button. So he was never able to go back to the layer edition, making
impossible to change a SQL view.

Fixed a small typo in a log while I was at it.