Adding ows-simulate community module.
When active allows to specify a "simulate" option to an OGC request
causing it to return after parsing the request but before actually
running it. The response to such a request is a serializastion of
the resulting "request object", which provides info about how the
request was parsed and interpreted.