[GEOS-7855] Fix for new style editor thinking old GeoServer CSS styles are SLD
Adds logic to the AbstractStylePage and StyleEditPage to check for and recover CSS styles generated by the old (pre-pluggable styles) CSS extension. If a CSS style is found, it is only recovered it if the derived SLD has not subsequently been edited.

(Squashed) Only recover CSS when current format is SLD; Better logging.