[GEOS-7944] Fix for null LegendInfo objects in StyleEditPage
When the AbstractStylePage is loaded with an existing style, it will now check if the StyleInfo object has a null LegendInfo. If so, it will insert a blank LegendInfo into the StyleInfo to act as the model for the ExternalGraphicPanel. This was already being done for new styles, but was missing for existing styles.

Added additional tests for discarding a legend on the StyleEditPage.

Fixed another related issue at the same time: Changed the “discard legend” functionality to fully reset user input — otherwise, supposedly discarded values could still fail to validate and prevent a save.