[Backport] [GEOS-7930] Prevent overwrite of updated styles when also changing workspace
Change the StyleEditPage submit function to first save the StyleInfo, then save the updated style body. When done in the opposite order, updates to the style body would be overwritten if the style was also moved.

Also added a test that isolates the bug by updating and moving a style at the same time.

Update: Merged test into existing StyleEditPageTests