GEOS-8033: Adding the "kmcentroid" option to KML.
The option provides the capability to specify options that control how
the placement of a KML placemark occurs. Namely to clip the geometry
before calculating the centroid, and the ability to generate a centroid
that falls within a polygon.

(cherry picked from commit b143eb0781e7045ace3f02540c833ad5fa8dc399)