Fix GeoGig CatalogVisitor so that it handles "autoIndexing" correctly.
When POSTing a GeoGig DataStore via the GeoServer REST API, the
"autoIndexing" connectionParameter value seems to be interpreted
as a String ("true"/"false") instead of a boolean (true/false).
This isn't the case when creating the DataStore via the GeoServer
UI. This change handles the value correctly in either case.

Signed-off-by: Erik Merkle <emerkle@boundlessgeo.com>