Move style POST examples to description
Move examples so that they show up properly
Swagger 2.0 does not support multiple examples/schemas for a single body object, which is neccessary for POST style
Instead, examples are in the description of the POST schema
Swagger Markdown does not properly support escaping XML (see https://github.com/swagger-api/swagger-ui/issues/3510 and https://github.com/swagger-api/swagger-ui/issues/3385 ). For now, using an image for the SLD example. This is awful, and we should change it as soon as Swagger Markdown is fixed.