Fix warning when parsing rules on a non global group
When parsing security rules, class DataAccessRuleDao log a warn when a
rule doesn't match an existing layer. This check doesn't handle
layergroups in workspaces, only workspaces, layers with or without
workspaces and layergroups without workspace.
This fix adds a check on layergroups in workspace in order to not log a
false warning.