Fix REST endpoint for datastore delete without recurse flag
Deleting a datastore without the recurse parameter set should
fail if there are FeatureTypes configured for that DataStore.

Instead, the logic was checking if the datastore existed
on its workspace, and failed if there's any other datastore
on the same workspace, which is unrelated to deleting a given
datastore.

This patch simplifies DataStoreController.dataStoreDelete()
letting the catalog comply if the DataStore cannot be deleted,
since the DataStore being empty is a precondition on
Catalog.remove(DataStoreInfo) anyways.

Removed the extra unnnecessary call to Catalog.remove().

Removed the calls to ResourcePool.clear() since that's already
handled by the internal machinery.