Add MapML preview of layer groups
Add (Java) support for > 1 feature in query responses, up to 50.

Catch and kill exception that occurs when feature extends outside of crs
projectable area e.g. north or south of ~85 degrees in Web Mercator.

Restrict what is returned by MapMLGetFeatureInfoOutputFormat to the first
feature of the last feature collection passed to it as a results list,
which is hopefully the 'top' feature.