Reduce resource usage of GWCInitializer
In order to register which tile layers don't have
to be cached by the in-memory cache, GWCIntializer
was doing two unnecessary things:

- loading all LayerInfo and LayerGroupInfo objects from
the GeoServer Catalog to an ArrayList. By one side, it's
unnecessarily resource consuming to load all catalog
layers and groups to a temporary list, and by the other
side, it is not necessary to load them at all, given
TileLayerCatalog can already provide the ids for the
configured tile layers from GeoServer layers and layergroups.

- synchronizing on the blobstore.memory.CacheProvider
while checking if a tilelayer has not to be in-memory
cached, inside a commpon-pool lambda. This is not needed
because CacheProvider is responsible for its own thread
safety and actually the one implementation, GuavaCacheProvider,
is already thread safe.