Proper handling of DefaultTileLayerCatalog life cycle
DefaultTileLayerCatalog was initializing itself if needed
whenever a public method was called, forcing unnecessary
synchronization on checkInitialized().

Given its life cycle should be managed by CatalogConfiguration,
this synchronization is not necessary.

The checkInitialized() method is repurposed to serve its
actual role of guarding public methods, and the life
cycle of DefaultTileLayerCatalog left to be managed
by the normal geoserver mechanisms.