web-gwc: String/RegExParameterFilterSubform workaround for ParameterFilter misbehavior
Thhis is a workaround for a side effect of an implementation detail in
RegexParameterFilter.getNormalize(), where when the value has not been
explicitly set, its instance variable remains null but the accessor returns a
new CaseNormalizer instance.

Before these GWC model objects correctly implemented equals() and hashCode(),
String/RefExParameterFilterSubform worked by accident, as a
by-product of CaseNormalizer.equals(Object) always evaluating to false when
confronted to any other instance.

Setting an explicit value for filter.normalize forces evaluation against
the actual instance and the FormComponent.convertInput() to behave as expected.

This workaround is side effect free since the net result of Filter having an
explicit value for its normalize instance variable is guaranteed anyways by
the way the form works.

In the long run what should happen is that
CaseNormalizingParameterFilter and its subclasses, as well as
CaseNormalizer, behave like simple POJOs instead of being clever about
returning a new default value on each accessor invocation, but that may
require further discussion at the GWC community level, which is beyond
the scope of this work, which is related to getting geoserver's gwc
integration working with the gwc configuration objects once they
correctly implement equals and hashCode, in the context of reducing
GWC configuration startup resource usage.