[GEOS-9458] Fix WFS 2.0 join queries with unqualified typenames
Improve org.geoserver.wfs.WFSWorkspaceQualifier to resolve
typenames on the local workspace when a WFS request with
multiple elements in the 'typeNames' (not 'typeName') parameter
are specified.

checkOriginallyUnqualified() was only looking at TYPENAME (WFS 1.0),
and not TYPENAMES (2.0).

Since by the time TypeNameKvpParser resolves the requested type names,
the LocalWorkspace is not set, it may either resolve to an unqualified
QName, or one on any other workspace than the requested one (due to the
lenient resolution in TypeNameKvpParser.parseToken() calling
Catalog.getFeatureTypeByName(String), which in turns returns a feature
type in the default workspace or any other workspace if none is found).
This lead to either trying to use the wrong FeatureType or not finding
it at all.

Also improve the readability of WFSWorkspaceQualifier.qualifyRequest(
 WorkspaceInfo, PublishedInfo, Service, Request)