Update win-installer.rst (#4599)
* Update win-installer.rst

The src folder is not in the bin distribution. Changed to src distribution. Also mentions that the user will have to download the src distribution.

* Update win-installer.rst

Removes reference to the GIT repository. REgular users would prefer to relate to the download website. This is where the user finds the bin and src-files which are used to compile an installer.