[GEOS-10371] Removing LayerGroup does not remove associated data acces rules
When a LayerGroup that belongs to a Workspace is removed,
its associated data access rules are removed.

When a root LayerGropup (i.e. null workspace) is removed,
its associated data access rules are not removed. Instead,
a NullPointerException is logged.

When a Workspace is removed, its data access rules are not
removed.

SecuredResourceNameChangeListener doesn’t check if a
LayerGroup.getWorkspace() is null, resuling in an NPE,
logged as a warning.

This patch simplifies the logic in SecuredResourceNameChangeListener that
would otherwise result in even more if-else clauses, makes sure
root layer group rules are treated properly, and adds test cases
for that, and for the rename cases, which weren't covered at all.