Update pom.xml to use CI-friendly versions through ${revistion} property
The project version is specified only in the root pom
as the "revision" property.

All child modules specify the parent's version as ${revision}.

This allows to change a single line in the root pom when
the project version has to be changed, instead of having
to modify all pom.xml files.

It also allows to specify the version is as command-line
argument (e.g. mvn -Drevision=2.21.0 clean install).

The flatten-maven-plugin resolves the revision variable and
installs all pom.xml files in the local maven repository with
the resolved value.

The sortpom-maven-plugin configuration is updated to set:
<pomFile>pom.xml</pomFile>
in order to format the actual pom instead of the intermediary
.flattened-pom.xml files created by flatten-maven-plugin.

See https://maven.apache.org/maven-ci-friendly.html