Link Loop
The link referring users to more details about the user_props table takes the user to a page where the table is never mentioned. The next page in sequence takes you back to this page. 

Test it out here:
https://docs.geoserver.org/latest/en/user/security/usergrouprole/usergroupservices.html

I'm reverse engineering how this is supposed to work, and will propose much better documentation regarding this when complete