[GEOS-12038] ModificationProxy.replaceCatalogInfo() precludes converting Filters to native query language
Use `rawCatalog` instead of `catalog` in ModificationProxy.replaceCatalogInfo()

ModificationProxy.replaceCatalogInfo() obtains the Catalog from GeoServerExtensions,
and is a private method used when cloning a CatalogInfo object. Using
the catalog bean implies going through the whole chain of Catalog decorators,
which includes AdvertisedCatalog, LocalWorkspaceCatalog, SecuredCatalog, etc.

The problem is it's called by ResourcePool every time a DataStore is requested,
then some CatalogInfo references need to be resolved by ResolvingProxy,
which calls the Catalog, and the decorators end up modifying the Filter used to do so,
which in turn precludes catalog backends (jdbcconfig/pgconfig) from optimizing the
query by id because the filter is too complex.