GEOS-10559 Lazy initialization of GetMapKvpRequestReader's HTTP client used to fetch remote styles
`GetMapKvpRequestReader` creates an apache
`CloseableHttpClient` on its constructor.

To do so, it needs to get the stles cache configuration from the
`org.geoserver.wms.WMS facade.`

Depending on the bootstrap conditions, this triggers an eager initialization
of beans, that can lead to a circular reference error.
This issue is observed in GeoServer Cloud.

Also, the eager initialization of the HTTP client can result in a waste
of resources, because it may never use it, and because there are subclasses
that definitely won’t use it.

Some of these subclasses are not spring beans, but created on-demand,
such as `WMSRequests.LayerParser`, which currently adds a
`ConfigurationListener` to `org.geoserver.config.GeoServer`
on each invocation of `WMSRequests.getGetMapParams()`, that is never
removed, hence being a memory leak.

As a solution, the HTTP client used by `GetMapKvpRequestReader` is now
lazily initialized, and the `ConfigurationListener` it uses to update
the cache configuration whenever `WMSInfo` is changed, removed whenever
the HTTP client is closed.