[GEOS-10560] Make GeoServerConfigurationLock reentrant
`GeoServerConfigurationLock` uses reentrant read/write locks
internally but does not preserve that behavior.

Since the lock can't know about the call chain that results
in a lock request, it's important that reentrancy is preserved.

A typical scenario is component A acquires a write lock to
perform a bunch of operations through component B, which also
acquires a write lock to perform more granular operations.

Component B unlocks, and then component A unlocks. Currently,
the `ThreadLocal<LockType>` is cleared on the first unlock,
leaving open locks when the next call to
`GeoServerConfigurationLock.unlock()` arrives.

Also, if a write lock is already held, and a read lock is requested
down the stream, the write lock shall be preserved.