[GEOS-10624] data directory ne workspace
This key addition is of the ne workspace containing a geopackage and associated layers for a simple world map. The geopackage is produced by a build/data/build.xml ant script that downloads and processes a geopackage using gdal/ogr.

This change updates contact information in support of [GSIP 202] Welcome Page:

* Welcome message introducing service
* For more information: both organization name and online resource to display a link
* Contact admin: requires email address mailto link