GEOS-10717: XStreamServiceLoader performance improvement with XstreamPersister caching
`XStreamServiceLoader` creates and initializes a new `XStreamPersister`
every time it has to save or load a ServiceInfo, which comes with
considerable overhead when depersisting thousands of service files.

Since `load()` receives a `GeoServer` argument, which has to be set
to the `XStreamPersiter`, a new `XStreamPersister` instance will
be created if `load()` is called with a different `GeoServer` reference.