[GEOS-10452] Fixed Role Services for LDAP/Active Directory.
Added the function "escapeSearchString()" to modify filter strings used for the Spring method template.search(...).
When having for example user CN's with a comma this comma is by default escaped with a "\". In the new Spring version the escape token "\" should be (double) escaped by "\\" to get proper results when searching on users etc.
Also an UnitTest is added.