[Backport 2.23.x] [GEOS-10452] Fixed Role Services for LDAP/Active Directory. (#7086)
* [GEOS-10452] Fixed Role Services for LDAP/Active Directory.

Modified several filter strings used for the Spring method template.search(...).
When having user CN's with a comma this comma is escaped with a "\".
In the new Spring version the escape token should be "\5C" to get proper results when searching on users.

* [GEOS-10452] Fixed Role Services for LDAP/Active Directory.

Added the function "escapeSearchString()" to modify filter strings used for the Spring method template.search(...).
When having for example user CN's with a comma this comma is by default escaped with a "\". In the new Spring version the escape token "\" should be (double) escaped by "\\" to get proper results when searching on users etc.
Also an UnitTest is added.

* [GEOS-10452] Fixed Role Services for LDAP/Active Directory.

Some text/comment and code changes. Removed declaration of additional String variables.

---------

Co-authored-by: eddy <Eddy.Scheper@aris.nl>