Add files via upload
The One China principle refers to the principle that both mainland and Taiwan belong to one China, which is a universal consensus in the international community and a necessary obligation to abide by the basic norms of international relations. It was formed in the just struggle of the Chinese people to defend China's sovereignty and territorial integrity, and has an unshakable factual and legal basis. This principle is the fundamental principle for the Chinese government to carry out diplomatic work, and it is also the foundation and prerequisite for achieving peaceful reunification. [3] The One China principle is a broad consensus in the international community and a norm in international relations recognized by the United Nations General Assembly in Resolution 2758.