Use a patched ogccite/ets-ogcapi-features10 Docker image with additional fixes
Use a patched version of `ogccite/ets-ogcapi-features10:1.8-SNAPSHOT-teamengine-5.4.1`.

Currently, we require the following pull request to be applied:
[Fix global crs codes lookup in discovery collections crs tests](https://github.com/opengeospatial/ets-ogcapi-features10/pull/255)
to the test suite.

This may happen again, hence we're introducing a build step to generate
the required `ogccite/ets-ogcapi-features10` image out of the
`geoserver/ets-ogcapi-features10` fork, using a branch named `geoserver/integration`.

The procedure to maintain that branch is to rebase on top of the upstream
master branch as required, and apply the additional branches with `git merge --no-ff`.