ogcapi-features CITE: build 1.8-teamengine-5.7 image out of the 1.8 tag upstream
Update `build/cite/ogcapi-features10/build-ets.sh` to use the new `1.8`
tag and hence the `ogccite/ets-ogcapi-features10:1.8-teamengine-5.7`
docker image, which is still not published in dockerhub.

This could be left as-is from now on since at least we now have a stable
git tag to build from that includes the required fixes to the test
engine.