[GEOS-11696] fix org.geoserver.sldservice.rest.RasterizerTest
`org.geoserver.sldservice.rest.RasterizerTest` lost the logged in user
set up at by the base class `CatalogRESTTestSupport`.

`getAsServletResponse(uri)` requests weren't failing per-se, as they
returned 200, but the layer wasn't found in `RasterizerController`.

`getAsDOM()` requests failed at parsing the response, but the underlying
reason was the same, the layer is not found since `AdminRequest` is set
but the user was `null`.

This patch removes the unnecessary override and refresh of the spring
application context, hence fixing the root cause of the problem.