[GEOS-11696] AdminRequestCallback not loaded due to spring bean name conflict
Rename spring bean `adminRequestCallback` to
`wicketAdminRequestCallback` in
`src/web/core/src/main/java/applicationContext.xml` to avoid component
name clash with restconfig's
`org.geoserver.rest.catalog.AdminRequestCallback`.

The later is loaded through component scan and ignored due to name
clashing when both gs-restconfig and gs-web-core are in the classpath.

---

`org.geoserver.sldservice.rest.RasterizerTest` lost the logged in user
set up at by the base class `CatalogRESTTestSupport`.

`getAsServletResponse(uri)` requests weren't failing per-se, as they
returned 200, but the layer wasn't found in `RasterizerController`.

`getAsDOM()` requests failed at parsing the response, but the underlying
reason was the same, the layer is not found since `AdminRequest` is set
but the user was `null`.

This patch removes the unnecessary override and refresh of the spring
application context, hence fixing the root cause of the problem.