[GEOS-11274,GEOS-11751] Generate URL for graphic inside workspace
GEOS-11274: In case of a single external graphic, use the workspace style directory to build the URL, since it is also published (similarly to the global style directory). This was causing a NPE ("iconUrl" is null).

GEOS-11751: When calling the icon service, use either the workspace-prefixed style name or add the workspace name to the path.