Fix `make war` in for build/cite
Add a install step before the package step to the build-war
Makefile target to avoid missing dependency errors for gs-css:test,
gs-mbstyle:test, and gs-app-schema-tests:test.

mvn --also-make does not conside test scoped dependencies, and with the
version change from 2.28-SNAPSHOT to 3.0-SNAPSHOT, those dependencies
can't be fetched from the maven repository as it's the first build with
that version.