[GEOS-11963] BlobStorePage breaks when failing to save a blob store configuration
Fix BlobStorePage breaking when failing to save a BlobStore

This is a long standing issue, when creating or modifying a BlobStore,
if there’s a problem saving it, the page breaks and presents a long stack trace.

The expected behavior is that it hows the error message and allows to
continue editing the blob store configuration parameters.

The root cause is that GWC.add/modifyBlobStore() declare to throw the
checked ConfigurationException, but it also throws the unckecked
ConfigurationPersistenceException. This later one was not considered.