[GEOS-11997] Migrate GeoServer ACL plugin to GeoServer codebase
Port the GeoServer ACL plugin from the geoserver-acl repository to src/community/acl

So far the GeoServer plugin source code was hosted in the
geoserver/geoserver-acl repository alongside the ACL libraries and
application as a measure to run tests in the CI build workflows.

This moves it to the GeoServer codebas as a community module in
anticipation of promoting it to the new pending profile and hence have
tests run as part of the regular GeoServer CI workflows.