[GEOS-10509] Fix support for XML requests where the first element is >8K or spans multiple read requests.
For XML POST requests, geoserver reads the first element to fetch some metadata, rewinds, and then parses the full XML. Rewinding currently happens using BufferedReader's mark support, but that requires specifying the maximum size of the first element.  As the first element can contain geometry and viewparams, it can be big.

This PR introduces a new class RewindableReader that allows rewinding with a dynamic buffer size.  To limit the size of the change, it gets wrapped with a BufferedReader and keeps the existing Request.setInput interface intact.

Geoserver requires the first call to Reader.ready() to return true, so RewindableReader supports it.

In theory, RewindableReader might replace all usage of XML_LOOKAHEAD.  However, it is unclear to me if the other cases are problematic or how to trigger the 8K edge for them, so this PR does only remove the case where GEOS-10509 demonstrates problems.

FIXUP