[GEOS-11926] ogcapi plugin makes WFS advertising an outputFormat which is actually unavailable
See https://osgeo-org.atlassian.net/browse/GEOS-11926 for the
motivation: when the RFCGeoJSONFeaturesResponse class is in the
classpath (e.g. when the ogcapi plugin is installed), then the class
will advertise itself as WFS outputformat in the WFS GetCapabilities,
but won't be actually usable when performing a WFS GetFeature.

This aims to fix the behaviour, by allowing the class to also be used by
the WFS subsystem.

Thanks Andrea Aime for the suggested fix.

Tests: added a utest to make sure we can use said class on a WFS
GetFeature operation.