removed redundant modal.close()
When viewing an elastic search layer, a modal can be triggered by clicking "Configure Elasticsearch fields" at the bottom of the screen. Closing this modal triggers a infinite loop when the close modal event listener attempts to close the modal which is already in the process of closing. Removing the line modal.close(target) resolves this issue.